Extract copy button into a shared component
