+++ Context/Context/Resources/ResourceDetailContent.swift
@@ -19,8 +19,6 @@ struct ResourceDetailContent: View {
   @State private var viewMode: ResourceViewMode = .preview
   @State private var rawResponseJSON: String? = nil
   @State private var requestError: (any Error)? = nil
-  @State private var showCopiedIndicator = false
-  @State private var copiedIndicatorTask: Task<Void, Never>?
 
   var body: some View {
     VStack(spacing: 0) {
@@ -171,22 +169,9 @@ struct ResourceDetailContent: View {
               
               // Copy button when in Raw mode
               if viewMode == .raw, rawResponseJSON != nil {
-                Button(action: {
+                CopyButton {
                   copyRawJSONToClipboard()
-                }) {
-                  HStack(spacing: 4) {
-                    Image(systemName: "doc.on.doc")
-                      .font(.system(size: 14))
-                    if showCopiedIndicator {
-                      Text("Copied!")
-                        .font(.caption)
-                        .foregroundColor(.secondary)
-                        .transition(.opacity.combined(with: .scale))
-                    }
-                  }
                 }
-                .buttonStyle(.plain)
-                .help("Copy to clipboard")
               }
               
               if isLoadingResources {
@@ -411,29 +396,5 @@ struct ResourceDetailContent: View {
     let unescapedJson = jsonString.replacingOccurrences(of: "\\/", with: "/")
     NSPasteboard.general.clearContents()
     NSPasteboard.general.setString(unescapedJson, forType: .string)
-    displayCopiedIndicator()
-  }
-  
-  private func displayCopiedIndicator() {
-    // Cancel any existing task
-    copiedIndicatorTask?.cancel()
-    
-    // Show indicator
-    withAnimation(.easeIn(duration: 0.1)) {
-      showCopiedIndicator = true
-    }
-    
-    // Hide indicator after delay
-    copiedIndicatorTask = Task {
-      try? await Task.sleep(nanoseconds: 1_500_000_000)  // 1.5 seconds
-      
-      if !Task.isCancelled {
-        await MainActor.run {
-          withAnimation(.easeOut(duration: 0.2)) {
-            showCopiedIndicator = false
-          }
-        }
-      }
-    }
   }
 }
