+++ Context/Context/Resources/ResourceTemplateDetailContent.swift
@@ -24,8 +24,6 @@ struct ResourceTemplateDetailContent: View {
   @State private var viewMode: ResourceViewMode = .preview
   @State private var rawResponseJSON: String? = nil
   @State private var requestError: (any Error)? = nil
-  @State private var showCopiedIndicator = false
-  @State private var copiedIndicatorTask: Task<Void, Never>?
 
   // Cached regex for template parameters
   private static let templateParameterRegex = /\{[^}]+\}/
@@ -245,22 +243,9 @@ struct ResourceTemplateDetailContent: View {
           HStack(spacing: 8) {
             // Copy button when in Raw mode with error
             if requestError != nil && viewMode == .raw && rawResponseJSON != nil {
-              Button(action: {
+              CopyButton {
                 copyRawJSONToClipboard()
-              }) {
-                HStack(spacing: 4) {
-                  Image(systemName: "doc.on.doc")
-                    .font(.system(size: 14))
-                  if showCopiedIndicator {
-                    Text("Copied!")
-                      .font(.caption)
-                      .foregroundColor(.secondary)
-                      .transition(.opacity.combined(with: .scale))
-                  }
-                }
               }
-              .buttonStyle(.plain)
-              .help("Copy to clipboard")
             }
             
             if isLoadingResources {
@@ -653,29 +638,5 @@ struct ResourceTemplateDetailContent: View {
     let unescapedJson = jsonString.replacingOccurrences(of: "\\/", with: "/")
     NSPasteboard.general.clearContents()
     NSPasteboard.general.setString(unescapedJson, forType: .string)
-    displayCopiedIndicator()
-  }
-  
-  private func displayCopiedIndicator() {
-    // Cancel any existing task
-    copiedIndicatorTask?.cancel()
-    
-    // Show indicator
-    withAnimation(.easeIn(duration: 0.1)) {
-      showCopiedIndicator = true
-    }
-    
-    // Hide indicator after delay
-    copiedIndicatorTask = Task {
-      try? await Task.sleep(nanoseconds: 1_500_000_000)  // 1.5 seconds
-      
-      if !Task.isCancelled {
-        await MainActor.run {
-          withAnimation(.easeOut(duration: 0.2)) {
-            showCopiedIndicator = false
-          }
-        }
-      }
-    }
   }
 }
